+++ cedar-policy/src/proto/ast.rs
@@ -20,6 +20,7 @@ use super::models;
 use cedar_policy_core::{
     ast, evaluator::RestrictedEvaluator, extensions::Extensions, FromNormalizedStr,
 };
+use itertools::Itertools;
 use smol_str::ToSmolStr;
 use std::{collections::HashSet, sync::Arc};
 
@@ -197,7 +198,7 @@ impl TryFrom<models::Entity> for ast::Entity {
             })
             .collect::<Result<Vec<_>, ProtobufConversionError>>()?;
 
-        Ok(Self::new_with_attr_partial_value(
+        Self::new_with_attr_partial_value(
             ast::EntityUID::try_from(
                 v.uid
                     .ok_or_else(|| ProtobufConversionError::missing("uid"))?,
@@ -206,7 +207,9 @@ impl TryFrom<models::Entity> for ast::Entity {
             HashSet::new(),
             ancestors,
             tags,
-        ))
+        )
+        .try_validate()
+        .map_err(|e| ProtobufConversionError::InvalidValue(format!("invalid entity: {e}")))
     }
 }
 
@@ -251,21 +254,21 @@ impl TryFrom<models::Expr> for ast::Expr {
             .expr_kind
             .ok_or_else(|| ProtobufConversionError::missing("expr_kind"))?;
 
-        match kind {
-            models::expr::ExprKind::Lit(lit) => Ok(ast::Expr::val(ast::Literal::try_from(lit)?)),
+        let expr = match kind {
+            models::expr::ExprKind::Lit(lit) => ast::Expr::val(ast::Literal::try_from(lit)?),
 
             models::expr::ExprKind::Var(var) => {
                 let pvar = models::expr::Var::try_from(var).map_err(|e| {
                     ProtobufConversionError::InvalidValue(format!("invalid var: {e}"))
                 })?;
-                Ok(ast::Expr::var(ast::Var::from(pvar)))
+                ast::Expr::var(ast::Var::from(pvar))
             }
 
             models::expr::ExprKind::Slot(slot) => {
                 let pslot = models::SlotId::try_from(slot).map_err(|e| {
                     ProtobufConversionError::InvalidValue(format!("invalid slot: {e}"))
                 })?;
-                Ok(ast::Expr::slot(ast::SlotId::from(pslot)))
+                ast::Expr::slot(ast::SlotId::from(pslot))
             }
 
             models::expr::ExprKind::If(msg) => {
@@ -278,11 +281,11 @@ impl TryFrom<models::Expr> for ast::Expr {
                 let else_expr = *msg
                     .else_expr
                     .ok_or_else(|| ProtobufConversionError::missing("else_expr"))?;
-                Ok(ast::Expr::ite(
+                ast::Expr::ite(
                     ast::Expr::try_from(test_expr)?,
                     ast::Expr::try_from(then_expr)?,
                     ast::Expr::try_from(else_expr)?,
-                ))
+                )
             }
 
             models::expr::ExprKind::And(msg) => {
@@ -292,10 +295,7 @@ impl TryFrom<models::Expr> for ast::Expr {
                 let right = *msg
                     .right
                     .ok_or_else(|| ProtobufConversionError::missing("right"))?;
-                Ok(ast::Expr::and(
-                    ast::Expr::try_from(left)?,
-                    ast::Expr::try_from(right)?,
-                ))
+                ast::Expr::and(ast::Expr::try_from(left)?, ast::Expr::try_from(right)?)
             }
 
             models::expr::ExprKind::Or(msg) => {
@@ -305,10 +305,7 @@ impl TryFrom<models::Expr> for ast::Expr {
                 let right = *msg
                     .right
                     .ok_or_else(|| ProtobufConversionError::missing("right"))?;
-                Ok(ast::Expr::or(
-                    ast::Expr::try_from(left)?,
-                    ast::Expr::try_from(right)?,
-                ))
+                ast::Expr::or(ast::Expr::try_from(left)?, ast::Expr::try_from(right)?)
             }
 
             models::expr::ExprKind::UApp(msg) => {
@@ -318,10 +315,7 @@ impl TryFrom<models::Expr> for ast::Expr {
                 let puop = models::expr::unary_app::Op::try_from(msg.op).map_err(|e| {
                     ProtobufConversionError::InvalidValue(format!("invalid unary op: {e}"))
                 })?;
-                Ok(ast::Expr::unary_app(
-                    ast::UnaryOp::from(puop),
-                    ast::Expr::try_from(arg)?,
-                ))
+                ast::Expr::unary_app(ast::UnaryOp::from(puop), ast::Expr::try_from(arg)?)
             }
 
             models::expr::ExprKind::BApp(msg) => {
@@ -334,14 +328,14 @@ impl TryFrom<models::Expr> for ast::Expr {
                 let right = *msg
                     .right
                     .ok_or_else(|| ProtobufConversionError::missing("right"))?;
-                Ok(ast::Expr::binary_app(
+                ast::Expr::binary_app(
                     ast::BinaryOp::from(pbop),
                     ast::Expr::try_from(left)?,
                     ast::Expr::try_from(right)?,
-                ))
+                )
             }
 
-            models::expr::ExprKind::ExtApp(msg) => Ok(ast::Expr::call_extension_fn(
+            models::expr::ExprKind::ExtApp(msg) => ast::Expr::call_extension_fn(
                 ast::Name::try_from(
                     msg.fn_name
                         .ok_or_else(|| ProtobufConversionError::missing("fn_name"))?,
@@ -350,60 +344,54 @@ impl TryFrom<models::Expr> for ast::Expr {
                     .into_iter()
                     .map(ast::Expr::try_from)
                     .collect::<Result<_, _>>()?,
-            )),
+            ),
 
             models::expr::ExprKind::GetAttr(msg) => {
                 let arg = *msg
                     .expr
                     .ok_or_else(|| ProtobufConversionError::missing("expr"))?;
-                Ok(ast::Expr::get_attr(
-                    ast::Expr::try_from(arg)?,
-                    msg.attr.into(),
-                ))
+                ast::Expr::get_attr(ast::Expr::try_from(arg)?, msg.attr.into())
             }
 
             models::expr::ExprKind::HasAttr(msg) => {
                 let arg = *msg
                     .expr
                     .ok_or_else(|| ProtobufConversionError::missing("expr"))?;
-                Ok(ast::Expr::has_attr(
-                    ast::Expr::try_from(arg)?,
-                    msg.attr.into(),
-                ))
+                ast::Expr::has_attr(ast::Expr::try_from(arg)?, msg.attr.into())
             }
 
             models::expr::ExprKind::Like(msg) => {
                 let arg = *msg
                     .expr
                     .ok_or_else(|| ProtobufConversionError::missing("expr"))?;
-                Ok(ast::Expr::like(
+                ast::Expr::like(
                     ast::Expr::try_from(arg)?,
                     msg.pattern
                         .into_iter()
                         .map(ast::PatternElem::try_from)
                         .collect::<Result<_, _>>()?,
-                ))
+                )
             }
 
             models::expr::ExprKind::Is(msg) => {
                 let arg = *msg
                     .expr
                     .ok_or_else(|| ProtobufConversionError::missing("expr"))?;
-                Ok(ast::Expr::is_entity_type(
+                ast::Expr::is_entity_type(
                     ast::Expr::try_from(arg)?,
                     ast::EntityType::try_from(
                         msg.entity_type
                             .ok_or_else(|| ProtobufConversionError::missing("entity_type"))?,
                     )?,
-                ))
+                )
             }
 
-            models::expr::ExprKind::Set(msg) => Ok(ast::Expr::set(
+            models::expr::ExprKind::Set(msg) => ast::Expr::set(
                 msg.elements
                     .into_iter()
                     .map(ast::Expr::try_from)
                     .collect::<Result<Vec<_>, _>>()?,
-            )),
+            ),
 
             models::expr::ExprKind::Record(msg) => {
                 let items = msg
@@ -413,9 +401,11 @@ impl TryFrom<models::Expr> for ast::Expr {
                     .collect::<Result<Vec<_>, ProtobufConversionError>>()?;
                 ast::Expr::record(items).map_err(|e| {
                     ProtobufConversionError::InvalidValue(format!("invalid record: {e}"))
-                })
+                })?
             }
-        }
+        };
+        expr.try_validate()
+            .map_err(|e| ProtobufConversionError::InvalidValue(format!("invalid expression: {e}")))
     }
 }
 
@@ -688,13 +678,11 @@ impl TryFrom<models::expr::like::PatternElem> for ast::PatternElem {
             .data
             .ok_or_else(|| ProtobufConversionError::missing("data"))?
         {
-            models::expr::like::pattern_elem::Data::C(c) => Ok(ast::PatternElem::Char(
-                c.chars().next().ok_or_else(|| {
-                    ProtobufConversionError::InvalidValue(
-                        "empty char in pattern element".to_string(),
-                    )
-                })?,
-            )),
+            models::expr::like::pattern_elem::Data::C(c) => {
+                Ok(ast::PatternElem::Char(c.chars().exactly_one().map_err(
+                    |e| ProtobufConversionError::InvalidValue(format!("{e} in pattern element")),
+                )?))
+            }
             models::expr::like::pattern_elem::Data::Wildcard(unit) => {
                 match models::expr::like::pattern_elem::Wildcard::try_from(unit).map_err(|e| {
                     ProtobufConversionError::InvalidValue(format!("invalid wildcard: {e}"))
@@ -1405,7 +1393,18 @@ mod test {
         };
         assert_matches!(
             ast::PatternElem::try_from(bad),
-            Err(ProtobufConversionError::InvalidValue(msg)) if msg.contains("empty char")
+            Err(ProtobufConversionError::InvalidValue(msg)) if msg.contains("got zero elements")
+        );
+    }
+
+    #[test]
+    fn pattern_elem_try_from_multi_char() {
+        let bad = models::expr::like::PatternElem {
+            data: Some(models::expr::like::pattern_elem::Data::C("foo".to_string())),
+        };
+        assert_matches!(
+            ast::PatternElem::try_from(bad),
+            Err(ProtobufConversionError::InvalidValue(msg)) if msg.contains("got at least 2 elements")
         );
     }