+++ cedar-policy/src/proto/policy.rs
@@ -101,7 +101,10 @@ impl From<&ast::Policy> for models::Policy {
 impl TryFrom<models::TemplateBody> for ast::Template {
     type Error = ProtobufConversionError;
     fn try_from(v: models::TemplateBody) -> Result<Self, Self::Error> {
-        Ok(ast::Template::from(ast::TemplateBody::try_from(v)?))
+        let template = ast::Template::from(ast::TemplateBody::try_from(v)?)
+            .try_validate()
+            .map_err(|e| ProtobufConversionError::InvalidValue(format!("invalid template: {e}")))?;
+        Ok(template)
     }
 }
 
@@ -487,6 +490,11 @@ impl TryFrom<models::PolicySet> for ast::PolicySet {
         let literal = ast::LiteralPolicySet::try_from(pset)?;
         ast::PolicySet::try_from(literal)
             .map_err(|e| ProtobufConversionError::InvalidValue(format!("invalid policy set: {e}")))
+            .and_then(|ps| {
+                ps.try_validate().map_err(|e| {
+                    ProtobufConversionError::InvalidValue(format!("invalid policy set: {e}"))
+                })
+            })
     }
 }