+++ cedar-policy-symcc/CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 Cedar Language Version: TBD
 
-## [0.5.1] - Coming soon
+## [0.5.1] - 2026-06-01
 Cedar Language Version: 4.5
 
 ### Fixed